Fix command typo in managing_daemons.diviner
Summary:
Fix the 'managing daemons' doc's example for launching daemons in a
way suitable for multiple web servers.  Separate the '--no-discovery'
argument with '--', otherwise it doesn't appear to be understood.

Test Plan:
Attempt to launch daemon in various ways, make sure to include
ways that are expected to fail, otherwise it may not be doing what
we expect.

    phd launch PhabricatorRepositoryPullLocalDaemon --no-discovery
    .. errors in daemon log 'unrecognised argument' ..
    phd launch PhabricatorRepositoryPullLocalDaemon -- --not-an-arg
    .. errors in daemon log 'unrecognised argument' ..
    phd launch PhabricatorRepositoryPullLocalDaemon -- --no-discovery
    .. no errors in daemon log ..

Reviewers: #blessed_reviewers, epriestley

Reviewed By: #blessed_reviewers, epriestley

Subscribers: epriestley

Differential Revision: https://secure.phabricator.com/D9374